requirements.txt
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-google-genai==2.0.8
python-dotenv
streamlit
unstructured==0.16.12
tiktoken==0.8.0
faiss-cpu==1.9.0.post1
libmagic==1.0
python-magic==0.4.27
python-magic-bin==0.4.14